+++ apps/api/test/routes/health.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import request from 'supertest';
-import app from '../src/app.js';
+import app from '../../src/app.js';
 
 describe('Health Endpoint', () => {
   it('should return success response with health data', async () => {
