+++ packages/shared/src/types/trains.ts
@@ -1,32 +1,3 @@
-// API response types
-
-/**
- * Generic API response wrapper
- */
-export interface ApiResponse<T = unknown> {
-  success: boolean;
-  data?: T;
-  error?: string;
-  message?: string;
-  timestamp: string;
-}
-
-/**
- * Success response wrapper
- */
-export interface SuccessResponse<T = unknown> extends ApiResponse<T> {
-  success: true;
-  data: T;
-}
-
-/**
- * Error response wrapper
- */
-export interface ErrorResponse extends ApiResponse<never> {
-  success: false;
-  error: string;
-}
-
 /**
  * Nearest trains request payload
  */
